pyproject.toml
[tool.poetry]
name = "nigerian-audit-ai"
version = "0.1.0"
description = "AI-powered audit system for Nigerian financial regulations."
authors = ["Jessie <jessyuba11@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
google-cloud-storage = "^2.10.0"
google-cloud-bigquery = "^3.23.0"
cloud-sql-python-connector = "^1.3.0"
google-cloud-logging = "^3.9.0"
google-cloud-monitoring = "^2.20.0"
google-cloud-aiplatform = "^1.62.0"
google-cloud-secret-manager = "^2.19.0"
SQLAlchemy = "^2.0.30"
psycopg2-binary = "^2.9.9"
pandas = "^2.2.2"
numpy = "^1.26.4"
scikit-learn = "^1.5.0"
tensorflow = "^2.16.1"
keras = "^3.3.3"
dvc = {extras = ["gdrive"], version = "^3.50.0"} # Or dvc = {extras = ["gs"], version = "^3.50.0"} for GCS
mlflow = "^2.13.0"
requests = "^2.32.3"
beautifulsoup4 = "^4.12.3"
lxml = "^5.2.2"
requests-html = "^0.10.0"
playwright = "^1.44.0"
great-expectations = "^0.18.10"
python-dotenv = "^1.0.1"
loguru = "^0.7.2"
tenacity = "^8.2.3"
typing-extensions = "^4.12.0"
pydantic = "^2.7.4"
pydantic-settings = "^2.3.3"
faiss-cpu = "^1.7.4"
sentence-transformers = "^2.2.2"
google-genai = "^1.0.0"
pytesseract = "^0.3.10"
Pillow = "^10.0.0"
pdfminer-six = "^20221105"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"